refactor: move unit tests to new module
